move tests related traces to create-trace.test.ts
