+++ src/traces/create-spans.ts
@@ -3,41 +3,11 @@ import {
   WorkflowRun,
   WorkflowRunJobs,
   WorkflowRunJob,
-  getLatestCompletedAt,
-  WorkflowResults
+  getLatestCompletedAt
 } from '../github/index.js'
 import * as opentelemetry from '@opentelemetry/api'
-import settings from '../settings.js'
-import * as core from '@actions/core'
 
-export const createTraces = async (
-  results: WorkflowResults
-): Promise<string | undefined> => {
-  if (settings.FeatureFlagTrace) {
-    console.log('trace feature is enabled.')
-  } else {
-    return undefined
-  }
-
-  const { workflowRun, workflowRunJobs } = results
-  const rootCtx = createWorkflowRunTrace(workflowRun, workflowRunJobs)
-  for (const job of workflowRunJobs) {
-    const jobCtx = createWorkflowRunJobSpan(rootCtx, job)
-    if (jobCtx === null) continue
-    createWorkflowRunStepSpan(jobCtx, job)
-  }
-
-  const traceId = opentelemetry.trace.getSpanContext(rootCtx)?.traceId
-  console.log(`TraceID: ${traceId}`)
-
-  // TODO: actions output traceID
-  // TODO: Delete this feature.
-  await createSummary(traceId)
-
-  return traceId
-}
-
-const createWorkflowRunTrace = (
+export const createWorkflowRunTrace = (
   workflowRun: WorkflowRun,
   workflowRunJobs: WorkflowRunJobs
 ): Context => {
@@ -53,7 +23,7 @@ const createWorkflowRunTrace = (
   return opentelemetry.trace.setSpan(ROOT_CONTEXT, span)
 }
 
-const createWorkflowRunJobSpan = (
+export const createWorkflowRunJobSpan = (
   ctx: Context,
   job: WorkflowRunJob
 ): Context | null => {
@@ -98,7 +68,10 @@ const createWorkflowRunJobSpan = (
   return opentelemetry.trace.setSpan(ctxWithWaiting, jobSpan)
 }
 
-const createWorkflowRunStepSpan = (ctx: Context, job: WorkflowRunJob): void => {
+export const createWorkflowRunStepSpan = (
+  ctx: Context,
+  job: WorkflowRunJob
+): void => {
   if (job.steps === undefined || job.steps.length === 0) {
     console.warn(`job (${job.name}) has no steps.`)
     return
@@ -139,16 +112,4 @@ const createSpan = (
   const span = tracer.startSpan(name, { startTime, attributes }, ctx)
   span.end(endTime)
   return span
-}
-
-const createSummary = async (traceId: string | undefined): Promise<void> => {
-  if (settings.isGitHubActions)
-    await core.summary
-      .addHeading('GitHub Actions OpenTelemetry')
-      .addRaw(`TraceID: ${traceId}\n`)
-      .addLink(
-        'Google Cloud Trace Helper',
-        `https://console.cloud.google.com/traces/list?tid=${traceId}`
-      )
-      .write()
 }
