+++ src/traces/index.ts
@@ -1,1 +1,1 @@
-export * from './github.js'
+export * from './create-trace.js'
